Preprint related fixes, newly collected data

+++ pipeline/prepper.py
@@ -1,10 +1,6 @@
-import multiprocessing
+from multiprocessing.pool import Pool
 import os
-import random
 import shutil
-import subprocess
-import time
-import re
 import pandas as pd
 
 DB_PATH = 'data'
@@ -13,8 +9,10 @@ DOT_PDB = '.pdb'
 HETATMS_DELETED = 'hetatms_deleted'
 
 PDB_PREP_DIR = 'pdb_prep'
-SCHROD_SCRIPT_PATH = 'resources/schrod_multi_prepare.sh'
-PDBFIXER_SCRIPT_PATH = 'resources/pdbfixer_multi_prepare.sh'
+# SCHROD_SCRIPT_PATH = 'resources/schrod_multi_prepare.sh'
+SCHROD_SCRIPT_PATH = 'resources/schrod_single_prepare.sh'
+# PDBFIXER_SCRIPT_PATH = 'resources/pdbfixer_multi_prepare.sh'
+PDBFIXER_SCRIPT_PATH = 'resources/pdbfixer_single_prepare.sh'
 
 PREP_SUFF = '.o.pdb'
 
@@ -29,14 +27,28 @@ DOT_FASTA = '.fasta'
 
 DB_INFO_PATH = 'db_info_{}.csv'
 
+NUMBER_OF_PROCESSES = 20
+# NUMBER_OF_PROCESSES = 1
+
+class RunPrep(object):
+  def __init__(self, name_to_path, tmp_dir, mode, last_epoch_name, epoch_name):
+      self.name_to_path    = name_to_path
+      self.tmp_dir         = tmp_dir
+      self.mode            = mode
+      self.last_epoch_name = last_epoch_name
+      self.epoch_name      = epoch_name
+  def __call__(self, name):
+      run_preparation_for_file(name, 
+        self.name_to_path, self.tmp_dir, self.mode, self.last_epoch_name, self.epoch_name)
 
 def get_pdb_paths(dir_path, prev_epoch, comps_to_take, fast_version=False):
     def accept_path(p):
         if not comps_to_take:
             return True
 
-        m = re.search('/(...._(.\+.|.)\|(.\+.\+.\+.|.\+.\+.|.\+.|.))/', p)
-        comp_name = m.group(1)
+        # m = re.search('/(...._(.\+.|.)\|(.\+.\+.\+.|.\+.\+.|.\+.|.))/', p)
+        # comp_name = m.group(1)
+        comp_name = os.path.basename(os.path.dirname(p))
 
         return comp_name in comps_to_take
 
@@ -53,7 +65,7 @@ def get_pdb_paths(dir_path, prev_epoch, comps_to_take, fast_version=False):
             if file.endswith(DOT_PDB):
                 pdb_paths.append(os.path.join(root, file))
 
-        random.shuffle(dirnames)
+        # random.shuffle(dirnames)
         candidate_dirs = dirnames if not fast_version else \
             dirnames[:min(2, len(dirnames))]
 
@@ -66,59 +78,66 @@ def get_pdb_paths(dir_path, prev_epoch, comps_to_take, fast_version=False):
 
     return pdb_paths
 
+def prep_in_mode(file_name, tmp_dir, mode):
+    expected_file = os.path.join(tmp_dir, file_name + PREP_SUFF)
 
-def await_expected_files(expected_files, tmp_dir):
-    res = {}
-
-    waiting_time = 10 * 60
-
-    while len(expected_files) > 0:
-        for path in os.listdir('.'):
-            if path.endswith(PREP_SUFF) and path in expected_files:
-                expected_files.remove(path)
-                name = path[:-len(PREP_SUFF)]
-                res[name] = os.path.join(tmp_dir, path)
-        time.sleep(1)
-        waiting_time -= 1
-
-        if waiting_time <= 0:
-            break
-
-    return res
-
-
-def prep_in_mode(file_names, tmp_dir, mode, arg=None):
-    expected_files = set(map(lambda x: x + PREP_SUFF, file_names))
+    if os.path.exists(expected_file):
+      return expected_file
 
     path_to_script = SCHROD_SCRIPT_PATH if mode == SCHROD else \
         PDBFIXER_SCRIPT_PATH
     script_name = os.path.basename(path_to_script)
 
-    shutil.copyfile(path_to_script,
-                    os.path.join(tmp_dir, script_name))
-    shutil.copymode(path_to_script,
-                    os.path.join(tmp_dir, script_name))
+    if not os.path.exists(os.path.join(tmp_dir, script_name)):
+      shutil.copyfile(path_to_script,
+                      os.path.join(tmp_dir, script_name))
+      shutil.copymode(path_to_script,
+                      os.path.join(tmp_dir, script_name))
 
     os.chdir(tmp_dir)
 
-    command = './{} {}'.format(script_name,
-                               arg) if arg else './{}'.format(script_name)
-    subprocess.call(command, stdout=subprocess.PIPE, shell=arg is not None)
+    command = f'./{script_name} {file_name}'
 
-    res = await_expected_files(expected_files, tmp_dir)
+    print(f'Running prepraration program... Command {command}')
+    status_code = os.system(command)
 
     os.chdir('..')
 
-    return res
+    if status_code != 0:
+      return None
+
+    return expected_file
 
 
-def schrod_prep(file_names, tmp_dir):
-    return prep_in_mode(file_names, tmp_dir, SCHROD, DB_PATH)
+def schrod_prep(file_name, tmp_dir):
+    return prep_in_mode(file_name, tmp_dir, SCHROD)
 
 
-def pdb_fixer_prep(file_names, tmp_dir):
-    return prep_in_mode(file_names, tmp_dir, PDBFIXER, DB_PATH)
+def pdb_fixer_prep(file_name, tmp_dir):
+    return prep_in_mode(file_name, tmp_dir, PDBFIXER)
 
+def run_preparation_for_file(name, name_to_path, tmp_dir, mode, last_epoch_name, epoch_name):
+  try:
+    new_path = name_to_path[name].replace(last_epoch_name, epoch_name)
+    if (os.path.exists(new_path)):
+      print(f'Prepared file already exists...')
+      return
+
+    path_to_prepped_schrod = schrod_prep(name, tmp_dir) if mode == SCHROD \
+        else pdb_fixer_prep(name, tmp_dir)
+
+    if path_to_prepped_schrod is None:
+      print(f'Preparation of {path_to_prepped_schrod} hasn\'t finished with 0 status code, skip it')
+      return
+
+    if not os.path.exists(os.path.dirname(new_path)):
+        os.makedirs(os.path.dirname(new_path))
+
+    print(f'Copy {path_to_prepped_schrod} to {new_path}...')
+    shutil.copyfile(path_to_prepped_schrod, new_path)
+  except Exception as e:
+    print(f'File {name}, ERROR:', e, flush=True)
+    return
 
 def prep_pdbs(last_epoch_name, epoch_name, db_path, mode, tmp_dir,
               fast_version, comps_to_take, run_id):
@@ -142,41 +161,25 @@ def prep_pdbs(last_epoch_name, epoch_name, db_path, mode, tmp_dir,
 
         name_to_path[name] = path
 
-        shutil.copyfile(path.replace(last_epoch_name, SEQS).
-                        replace(DOT_PDB, DOT_FASTA),
-                        os.path.join(tmp_dir, name) + DOT_FASTA)
-        shutil.copyfile(path, os.path.join(tmp_dir, name))
+        if not os.path.exists(os.path.join(tmp_dir, name) + DOT_FASTA):
+          shutil.copyfile(path.replace(last_epoch_name, SEQS).replace(DOT_PDB, DOT_FASTA),
+                          os.path.join(tmp_dir, name) + DOT_FASTA)
+        if not os.path.exists(os.path.join(tmp_dir, name)):
+          shutil.copyfile(path, os.path.join(tmp_dir, name))
 
-    unprepped_names = set(name_to_path.keys())
+    unprepped_names = list(set(name_to_path.keys()))
 
-    while len(unprepped_names) > 0:
-        print('New iteration:', len(unprepped_names), flush=True)
+    print(f'len(unprepped_names)={len(unprepped_names)}')
 
-        with open('unprepped_{}_{}.log'.format(mode, run_id), 'w') as f:
-            for name in unprepped_names:
-                f.write(name_to_path[name] + '\n')
-                f.flush()
-
-        path_to_prepped_schrod = schrod_prep(unprepped_names,
-                                             tmp_dir) if mode == SCHROD \
-            else pdb_fixer_prep(unprepped_names, tmp_dir)
-
-        for name in list(unprepped_names):
-            if name not in path_to_prepped_schrod.keys():
-                continue
-
-            unprepped_names.remove(name)
-
-            path_to_prepped = path_to_prepped_schrod[name]
-
-            new_path = name_to_path[name].replace(last_epoch_name, epoch_name)
-
-            if not os.path.exists(os.path.dirname(new_path)):
-                os.makedirs(os.path.dirname(new_path))
-
-            shutil.copyfile(path_to_prepped,
-                            new_path)
-
+    if NUMBER_OF_PROCESSES == 1:
+      print(f'Running preparation sequencially...')
+      for name in unprepped_names:
+        run_preparation_for_file(name, name_to_path, tmp_dir, mode, last_epoch_name, epoch_name)
+    else:
+      print(f'Running preparation in parallel with {NUMBER_OF_PROCESSES} processes...')
+      with Pool(NUMBER_OF_PROCESSES) as pool:
+        run_prep = RunPrep(name_to_path, tmp_dir, mode, last_epoch_name, epoch_name)
+        pool.map(run_prep, unprepped_names)
 
 if __name__ == '__main__':
     from optparse import OptionParser
@@ -218,11 +221,20 @@ if __name__ == '__main__':
         is_fast_version = False
 
     comps_to_take = set()
+    
+    print(f'Reading csv {DB_INFO_PATH.format(options.run_id)}...')
+
     db_info = pd.read_csv(DB_INFO_PATH.format(options.run_id))
+    
+    print(f'csv read')
 
     for i in range(len(db_info)):
         comps_to_take.add(db_info.iloc[i]['comp_name'])
 
+    print(f'Preparing structures...')
+
     prep_pdbs(options.prev_epoch, options.cur_epoch, options.db, options.mode,
               options.tmp_dir + '_' + options.run_id,
               is_fast_version, comps_to_take, options.run_id)
+
+    print(f'Everything is finished!...')
